% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_results.R
\name{final_results}
\alias{final_results}
\title{Write results}
\usage{
final_results(params, write)
}
\arguments{
\item{params}{parameters from mcmc}

\item{write}{logical value of saving results}
}
\value{
results
}
\description{
Write results from mcmc
}
\examples{
example(final_results(params, write = F))

}
