% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_data.R
\name{generate_data}
\alias{generate_data}
\title{Generate data}
\usage{
generate_data(
  n,
  theta10,
  theta11,
  thetaT,
  alpha1,
  psi2,
  beta0,
  omega1,
  beta1,
  omega2,
  eps1,
  eps2,
  eps3,
  sigmae,
  misspecify,
  r
)
}
\arguments{
\item{n}{number of observations to simulate}

\item{theta10}{correlation between S(1) and T(0)}

\item{theta11}{correlation between S(1) and T(1)}

\item{thetaT}{correlation between T(0) and T(1)}

\item{alpha1}{mean of S(1)}

\item{psi2}{effect of X on S(1)}

\item{beta0}{mean of T(0)}

\item{omega1}{effect of X on T(0)}

\item{beta1}{mean of T(1)}

\item{omega2}{effect of X on T(1)}

\item{eps1}{standard deviation of S(1)}

\item{eps2}{standard deviation of b0}

\item{eps3}{standard deviation of b1}

\item{sigmae}{standard deviation of residual error}

\item{misspecify}{logical value if mvn model is misspecified}

\item{number}{of repeated measures}
}
\value{
dataset
}
\description{
Generate data for simulations for random intercept model
}
\examples{
example(generate_data(n = 100, theta10 = 0.1, theta11 = 0.3, thetaT = 0.7,
 alpha1 = 1, psi2 = 1,
 beta0 = 2, omega1 = 3,
 beta1 = 1, omega2 = 3,
 eps1 = 1, eps2 = 1, eps3 = 1,
 sigmae = 0.3,
 misspecify = FALSE, r = 2))

}
