% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.R
\name{run_mcmc}
\alias{run_mcmc}
\title{Run MCMC}
\usage{
run_mcmc(
  samp,
  Xi,
  n,
  SIM,
  r,
  condindfit,
  prior,
  r13init,
  r12init,
  r23init,
  b0init,
  b1init
)
}
\arguments{
\item{samp}{data}

\item{Xi}{covariates}

\item{n}{sample size}

\item{SIM}{number of mcmc iterations}

\item{r}{number of repeated measures}

\item{condindfit}{logical value to assume conditional independence}

\item{prior}{either uniform or beta prior on theta}

\item{r13init}{initial value of theta11}

\item{r12init}{initial value of theta10}

\item{r23init}{initial value of thetaT}
}
\value{
a list of samples
}
\description{
Run MCMC ...
}
\examples{
example(run_mcmc(samp = samp, Xi = rep(0, n),  n = n, SIM = SIM, r = r, condindfit = condindfit, 
prior = prior, r13init = 0.6, r12init = 0.1, r23init = 0.1,
b0init = rep(0, n/2), b1init = rep(0, n/2)))
}
